Allow fixed custom size for QTerminal
Closes https://github.com/lxqt/qterminal/issues/492

If the window size is not remembered on exiting, the custom width and height could be set in Preferences. Then, QTerminal will always have that size on starting. Of course, the maximized state is also saved/restored, as before.

Although the size is in pixels (not in characters), a button is added for getting the current size, so that the user could first resize QTerminal and then open Preferences and get the size. In this way, (1) lots of complex calculations are prevented, and (2) differences in tastes are respected.